requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
torch==2.5.1
transformers==4.48.2
diffusers==0.32.2
pillow==10.4.0
base64